Add dynamic commands for editing settings
Sublime Text 3 supported plugin-folder-relative paths for the
edit_settings command's base_file argument. ST4 doesn't (yet)
support this, so I've created two plugin commands to edit the
settings files using dynamically generated paths pointing to
the plugin's directory.

The path must be dynamic because SublimeGremlins was renamed
to "Gremlins" for packagecontrol.io, so the plugin might live
in a different path depending how it was installed.

Thanks to @luckman212 for flagging the ST4 issue in #10